fixup! cmake: multi image build support
We need to cache the list of images to validate the --image argument
in the runner scripts.

Also cache build strategy for each image -- only the default strategy
will work as-is. The hex file strategy can be supported with
--hex-file, but I'd rather not keep building on that, so let's defer
until we have a -f / --file argument guarded by a runner
capability (and we can sanity check this). It's not clear what to do
for the "skip build" strategy; does it even make sense? Need to
specify the use cases more carefully here IMO.

Signed-off-by: Marti Bolivar <marti.bolivar@nordicsemi.no>