boards: nrf52_sparkfun: fix board.cmake
This can't possibly have worked since the initial merge of the board.

It looks like it was originally written as a port from a KBuild
Makefile fragment but never tested, and has only been touched by
tree-wide changes since then. Try to fix it.

I don't have this hardware, but it should work the same way as
96b_nitrogen if it truly supports pyocd.

Signed-off-by: Marti Bolivar <marti.bolivar@nordicsemi.no>