drivers: led_strip: modernize ws2812 drivers and sample
Convert the GPIO based driver to the new GPIO API. (Only the
gpio_configure() call is affected).

Move configuration to DT where appropriate for both SPI and GPIO
drivers, only leaving the SPI vs. GPIO decision in Kconfig (in
addition to the basic enable for the driver.) Move some files around
to clean up as a result of this change.

Change the sample around as well to be a little easier to look at and
rely on DT.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>