samples: bme280: nrf52840 DK: set LSB in i2c addr
The nrf52840-DK doesn't have this sensor built-in, so it's likely
users will be getting this sensor from a breakout board.

The ones that Sparkfun and Adafruit sell pull up the LSB address pin,
so the default address is 0x77 on those boards. Let's follow along to
make this sample work with the boards that come up first in the
results when you google "BME280 breakout board" at time of writing.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>