drivers: gpio: add CONFIG_GPIO_PIN_OPERATIONS
No functional changes expected.

This option will be used in later commits to extend the GPIO driver
API to support "pin-based" GPIO controllers. After the API is ready,
Zephyr will be able to support a wider variety of GPIO controllers
than it is presently able to.

The cost of this additional support is increased footprint and runtime
penalties, so the option is disabled by default, and is meant to stay
disabled unless at least one GPIO driver in the system needs it.

The term "pin-based" is used here to refer to GPIO controller hardware
that manages large numbers of pins, typically 100-300 pins per
controller.

Pin-based controllers are common in "larger" SoCs. Each such controller
manages hundreds of pins, at the cost of:

    - significantly larger register block footprints in physical
      memory, and

    - the loss of the ability to do things like get all the pin states
      in the controller at once with a single register read

The workarounds I am aware of for managing such controllers in the
Zephyr GPIO API are:

    1. Artificially split up the controller into multiple GPIO
       devices, with each handling 32 of the pins managed by
       the overall controller

    2. Allocate an entire struct device per GPIO

Both approaches have obvious comparisons to Greek mythology's "bed of
Procrustes".

This is necessary because Zephyr's GPIO API's concept of a
"gpio_port_pins_t" is designed around a small, fixed number of pins
per GPIO controller.

The register interfaces for pin-based controllers really are a single
peripheral with some common state, plus hundreds of registers for
controlling individual pins. The size-limited concept of a small
number of pins per port is an implementation detail in the Zephyr API
that does not work well with the reality of the hardware, and we can
do better by accounting for both types of controller in the API.

Signed-off-by: Martí Bolívar <marti.bolivar@oss.qualcomm.com>