devicetree: edtlib: add helpers for gpio-reserved-ranges
This property's semantics are defined in dt-schema. Dealing with it
purely in C and the Zephyr bindings language yields suboptimal
results, with implementation-defined limits in terms of the number of
reserved ranges and the error-checking related to the value of the
property itself.

It will therefore enable more robust behavior and better error
messages if we have a first-class abstraction for this in edtlib. Add
one so we can use it in other layers.

Signed-off-by: Martí Bolívar <marti.bolivar@oss.qualcomm.com>