doc: west: document west.commands.WestCommand.die_if_no_git()
This method was added in west commit 7983cba0f0d134c45f8cfbab6b59aafa59c9367c
("WestCommand: lift die_if_no_git from project.py"), which was added
after v0.10.0 but before v0.11.0.

Document it since we are getting a user request that we expose it as
public API, and there's no reason not to.

Signed-off-by: Martí Bolívar <marti.bolivar@oss.qualcomm.com>